:recycle: Rename file name and move public function to XChanger.swift
